Fixing ADM-194 (Restore doesn't work in zarafabackup module)
The problem was, that the JSON request data was encoded twice. This
happened, because the zarafabackup code explicitly called escape()
on the data, although the method submitURL, which is used for sending
the form data, already does escaping by itself.

This bug was probably introduced because the zarafabackup code was
written before submitURL gained escaping support (in 79a5701 for
ADM-174).