Fixing ADM-118 (Missing error message if perl fails)
If the perl backend encounters an uncaught error, it might silently
be ignored by the UI. This is clearly a problem as the user might
think everything went well.

Normally, errors should be detected by the AJAX handler which calls our
error handler on failure (theme-core/javascript/ui: handleFailure).
This only works, if the HTTP status code of the reply is non-2xx,
as all 2xx replies are considered a success.
In reality there are circumstances where the HTTP headers have already
been printed (and the status code is 200 as such), but a later call
to some Perl or library function fails. In these cases, the error
message is hidden somewhere in the resulting HTML code.

This commit improves the checkError method in Yaffas.UI to look out
for such error messages and display them to the user.

Successful requests, requests failing regularly in Yaffas and (this is
the new feature) requests which fail unexpectedly due to Perl errors
mid-request have been tested to be handled properly.