Fixing ADM-155 (New first use wizard will give an unclear error when creating a user that already exists.)
The first use wizard blindly tried to create the non-privileged user,
without checking if the user exists locally already.

We now do a check just before the setup wizard performs the
real actions on the system. There is still a small window for a race
condition, but this should be highly unlikely and if it happens, the
only negative side effect is a sub-optimal error message.