Fixing ADM-196 ("Delete user" feature is (at least partly) broken) Part 1/2
The delete user function did not properly show the user name in the
confirmation dialog.
It also transmitted "X" instead of the user name to the
check_deluser.cgi script, which did not matter, as check_deluser.cgi
resolved he uid to the username by itself, ignoring the username
argument. So this part has just been dropped.