Avoid losing POST data when using ui.submitForm() in combination with module.confirmation() hooks which do their own AJAX request.
E.g.:
- normal form gets submitted, submitForm() hooks in, pre-processed
  stuff
- submitForm() calls submitURL() which itself calls
  module.confirmation() to allow interception by the module
- module.confirmation() sends another AJAX request
- the anonymous submit function (generated in submitURL, passed to
  confirmation() as third arg) is called
- post data is lost (empty POST request is sent to the server)

This fix should be unobtrusive. It calls .setForm() as it used to,
but also calls it immediately before sending the AJAX request,
so it is restored properly if it has been reset meanwhile.
So the worst behaviour change this commit may cause is that some
CGI script might get unexpected additional input. But even this
case requires a rather obscure calling chain...

This fix is needed for the implementation of ADM-202.