Fixing fetchmail's save poll to handle exceptions in a user-friendly way
save_poll.cgi lacked a try { } catch {}; block, so any exception
thrown would lead to a perl error instead of a nicely formatted
error message.

This commit fix this in a huge hack (stuffing all the existing code
into a single function and wrapping a call to this function in
try/catch)

Also for ADM-206.