yaffas-config: Allow upgrade scripts to use Bash features
We require bash anyway (for the upgrade script itself), partly make
use of bash features (such as $()), but execute the upgrade scripts
as "sh" (which at least on Debian-based systems leads to a very
restricted shell feature set).
We're changing this to bash now, so that we can use the faster and
safer [[ ]] tests and other features.