Fixing ADM-215 (responses to file upload requests are not properly handled by the UI)
Our UI response parsing code was wrong, although it worked in most
cases. We are using the responseText property of YUI2's async
response objects. This apparently works fine for AJAX/XHR requests,
but it breaks whenever iframes are used as the request method
(YUI2 transparently does that for file uploads).
Where "breaks" means: .responseText contains only the text (all HTML
stripped) in these cases (as documented...).
The correct way to go is to use .responseXML. Sadly, this is not always
populated.
So the fix is: We try to use .responseXML, if available, but fall back
to .responseText in all other cases.
This seems to work fine on Firefox and Chrome for all types of
requests (GET, POST, POST w/ file upload).

As .responseText was used multiple times in ui.js, I moved this
decision logic to its own function, which is now used instead.
(To avoid code duplication)