Fixing ADM-217 (Info message is bigger thatn screen without scroll bar. Unable to see and click buttons.)
Our confirmation dialogs (as used in Zarafa -> Configuration when
asking for permission to delete users when changing AD filter) did
not have any maximum size. Depending on the size of the dialog
contents and the available space (resolution / window size dependent),
this was a problem, because the Yes/No buttons were no longer
clickable.

We've fixed this by calculating a sane max-height (using JS) for the
dialog body and setting overflow: auto, so that scrollbars are added
as necessary by the browser.
We also monitor the window for size changes so that the dialog
can be expanded/shrinked as necessary.

Tested on Firefox 16/Linux, Chromium 20/Linux, IE9/Windows 7
Needs testing in different browsers.