Workaround for ADM-235 (Username lookup sometimes broken (e.g. when attempting to modify a user)
We change the get_username_by_uid implementation from a
(strangely wrapped) getpwuid() call which is known to cause the
mentioned bug to an implementation which uses our own
getent("passwd") call/cache, as all the other functions in the module
do already.
The code is borrowed from get_groupname_by_gid.
Yaffas::UGM::gecos might still cause problems if the getpw* functions
fail, as it is still using getpwnam()