SNMP config improvements as per ADM-226, ADM-224
This commit introduces some important changes to the SNMP config
module.
We previously tried to patch the existing snmpd.confs, which is no
longer feasable because of distribution-specific changes.

We now replace the existing config with a yaffas-provided, standardized
file and adjust this one as needed. The old config is moved to
.yaffassave.

This fixes
  - ADM-226 (SNMP is broken on Ubuntu 12.04 and Debian 6)
  - SNMP configuration produces totally different behaviour depending on the distribution