yaffas-core: Fixing build on RHEL
yaffas-core failed to build on RHEL unless the git checkout was clean.
We now enforce this in make clean and call it before patching in the
build phase.
Proper solution would be to copy the whole usr/, etc/ and var/ dirs
somewhere to the build dir and patch them there...