Fixing ADM-257 (Config restore fails without error message and leaves broken LDAP behind)
During configuration restoration domrename.pl is called to fix
differences in the LDAP scheme between the previously set domain
and the new domain.
domrename.pl's fix consists of doing some search-and-replace magic.

The root dn entry is a special case and ends up broken after the
search-and-replace magic. Therefore, domrename.pl removes the first
block it encounters and re-adds a well-formed version.

The remove-first-block logic was broken and this became apparent
whenever the input LDIF contained a leading new line. In all these
cases the broken root dn block was not properly removed, but the
well-formed block was added nevertheless, resulting in two root dn
blocks.

The later call to slapadd using this LDIF file then lead to an error
due to the duplicate insertion attempt of the root dn.
This led to slapcat exiting with a non-zero exit code, even if the
rest of the LDIF was properly imported.

The configuration restoration routine therefore considered the LDAP
restoration to be failed and did not proceed with the remaining LDAP
update code such as correcting the permissions of /var/lib/ldap
and restarting several services.

Note: This error lead to slapd being broken, but no error message was
ever being displayed to the user due to ADM-215.