Unifying webmin restart calls (fixing ADM-268)
ADM-268: Config restore is executed multiple times and might lead
to an unpredictable system state

These changes drop all the fork()ing code used in different places to
execute webmin restarts. Instead, we handle this special case
directly in Yaffas::Service::control(WEBMIN, RESTART), so that the
caller does not have to care about it.

Instead of forking, we opt for an easier and less problematic method:
we simply ignore SIGTERMs for the currently active connection which
starts the webmin restart init script call. Therefore it is protected
from being killed and the control flow will continue as expected
afterwards.

No more stray children or code running in parallel although it
shouldn't.