Fixing ADM-281 (Upgrade from 1.2.3 to 1.3.x leaves old yaffas processes behind)
This is a CentOS/RHEL-only issue. The problem is that yaffas-core
unconditionally calls /etc/init.d/yaffas start in post-inst. As the
init script always starts a new process, even if an old process is
still running, we end up with two (or more) processes.
To avoid that, we attempt to stop yaffas first (but fail gracefully,
in contrary to the init script's restart option, if no yaffas process
is running (i.e. first install)).