Fixing ADM-288 (Config restore broken (LDAP/users))
Config restores were broken (slapadd failed, leading to an empty LDAP tree)
- if we were on RHEL/CentOS (SELinux) and
- if old and new domain matched

This got broken by the fix for ADM-265.

The problem is that for config restores where the old domain is identical
to the new domain, domrename.pl is no longer called. But domrename.pl does
more than just patching the LDIF, it also sets the correct SELinux context.
In other words: This did not happen (but is needed on CentOS/RHEL) if the
domains are unchanged.

Now, we ensure that the SELinux context is set even if domrename.pl is not
called.