Fixing ADM-294 (Changing the workgroup breaks network configuration until next reboot)
The cause for this problem is that Debian defaults to
"allow-hotplug eth0" in /etc/network/interfaces instead of
"auto eth0". This leads to ifdown -a turning the interface down, but
ifup -a not turning it up again (as it only considers
"auto" type interfaces there).

The fix consists of running ifup on all supported interfaces
explicitly.