Fixing ADM-273 (Not possible to create users with special characters in their GivenName or Surname)
This is the first step implementing this and consists of replacing
the gecos check "printable ASCII or German umlauts" with a check
which only blacklists some ASCII control characters and permits
all other unicode chars.
Also, we used to convert German umlauts to ASCII equivalents, this
function has also been removed.