Fixing ADM-311 (Group deletion broken in Chrome)
All data tables in yaffas which relied on a specific table cell
value of a selected row were broken in Chrome. This is related to
our Yaffas.Table javascript code, which first attempts to use
.innerText of the cell. .innerText contains the value with a
trailing new line in Chrome, while it does not in Firefox.
I have changed this code to use .textContent if available first and
only fall back to .innerText.

Not yet tested in IE, but unless it has a broken .textContent value
it should work...