Proper fix for ADM-304 (WebApp password change plugin not enabled when upgrading from pre-1.3)
The previous fix did not work. We tried to use perl as the interpreter in the
shebang line, which cannot work as the upgrade scripts are called with
"bash /path/to/script". So this is now a bash script with a single call to
perl.
Also, a perl module name was wrong.