Fixing root password changing
Did not work on some (!?) systems due to a fork being used, which
lead to UGM::password() to fail, which itself started "expect" and
lacked a tty.
The fork was supposed to protect against webmin restarts, but this
is already done by a TERM signal handler.

This is a bitkit-only thing.