yaffas-module-certificates: removing dependency on yaffas-certificates
Doing this to remove a cyclic dependency between these two which leads
to a wrong order of installation at least on RHEL.
The source of this issue is that yaffas-certificates used to provide
files, but now uses code from yaffas-module-certificates to generate
certificates on the fly.

Technically yaffas-module-certificates does not depend on
yaffas-certificates, so this is not a problem.