mailsrv: stop blindly replacing virtual_alias_maps in postfix's main.cf
The relevant function toggle_distribution_groups() is called from
AuthSrv as well, meaning it will always overwrite our default
config's virtual_alias_maps from yaffas-postfix upon installation.

While this could be handled in yaffas, this is certainly problematic
with code paths in Bitkit which need to change this value.

What we do now is simply replacing the relevant entry and leaving
everything else as-is.